+++ .flake8
@@ -1,2 +1,2 @@
 [flake8]
-ignore = E501
+ignore = E402, E501, F401
