+++ Pipfile
@@ -4,15 +4,61 @@ verify_ssl = true
 name = "pypi"
 
 [packages]
+alembic = "==1.8.1"
+asttokens = "==2.0.8"
+backcall = "==0.2.0"
+click = "==8.1.3"
+decorator = "==5.1.1"
+dnspython = "==2.2.1"
+dominate = "==2.7.0"
+email-validator = "==1.2.1"
+executing = "==1.0.0"
+flake8 = "==5.0.4"
 flask = "==2.0.2"
 flask-bootstrap = "==3.3.7.1"
+flask-login = "==0.6.2"
+flask-migrate = "==3.1.0"
+flask-sqlalchemy = "==2.5.1"
 flask-wtf = "==1.0.0"
+greenlet = "==1.1.3"
 gunicorn = "==20.1.0"
-ipython = "*"
-ipython-genutils = "*"
+idna = "==3.3"
+importlib-metadata = "==4.12.0"
+ipython = "==8.5.0"
+ipython-genutils = "==0.2.0"
+itsdangerous = "==2.1.2"
+jedi = "==0.18.1"
+jinja2 = "==3.1.2"
+mako = "==1.2.2"
+markupsafe = "==2.1.1"
+matplotlib-inline = "==0.1.6"
+mccabe = "==0.7.0"
+parso = "==0.8.3"
+pexpect = "==4.8.0"
+pickleshare = "==0.7.5"
 piecash = "==1.2.0"
-PyYAML = "==6.0"
-voluptuous = "*"
+prompt-toolkit = "==3.0.31"
+ptyprocess = "==0.7.0"
+pure-eval = "==0.2.2"
+pycodestyle = "==2.9.1"
+pyflakes = "==2.5.0"
+pygments = "==2.13.0"
+pytz = "==2022.2.1"
+pytz-deprecation-shim = "==0.1.0.post0"
+pyyaml = "==6.0"
+six = "==1.16.0"
+sqlalchemy = "==1.3.24"
+sqlalchemy-utils = "==0.38.3"
+stack-data = "==0.5.0"
+traitlets = "==5.4.0"
+tzdata = "==2022.2"
+tzlocal = "==4.2"
+visitor = "==0.1.3"
+voluptuous = "==0.13.1"
+wcwidth = "==0.2.5"
+werkzeug = "==2.2.2"
+wtforms = "==3.0.1"
+zipp = "==3.8.1"
 
 [dev-packages]
