+++ app/__init__.py
@@ -1,0 +1,11 @@
+from os import environ as env
+from flask import Flask
+from flask_bootstrap import Bootstrap
+
+
+app = Flask(__name__)
+app.config['SECRET_KEY'] = env.get('FLASK_SECRET_KEY',
+                                   'Mjpe[){i>"r3}]Fm+-{7#,m}qFtf!w)T')
+bootstrap = Bootstrap(app)
+
+from app import routes
