+++ app/static/delete_txn.css
