+++ app/static/easy-buttons.js
