+++ app/templates/delete_txn.html
