+++ app/templates/easy-buttons.html
