+++ requirements.txt
@@ -1,49 +1,55 @@
-#
-# These requirements were autogenerated by pipenv
-# To regenerate from the project's Pipfile, run:
-#
-#    pipenv lock --requirements
-#
-
--i https://pypi.org/simple
+alembic==1.8.1
 asttokens==2.0.8
 backcall==0.2.0
-click==8.1.3; python_version >= '3.7'
-decorator==5.1.1; python_version >= '3.5'
-dominate==2.7.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
+click==8.1.3
+decorator==5.1.1
+dnspython==2.2.1
+dominate==2.7.0
+email-validator==1.2.1
 executing==1.0.0
-flask-bootstrap==3.3.7.1
-flask-wtf==1.0.0
-flask==2.0.2
+flake8==5.0.4
+Flask==2.0.2
+Flask-Bootstrap==3.3.7.1
+Flask-Login==0.6.2
+Flask-Migrate==3.1.0
+Flask-SQLAlchemy==2.5.1
+Flask-WTF==1.0.0
+greenlet==1.1.3
 gunicorn==20.1.0
-ipython-genutils==0.2.0
+idna==3.3
+importlib-metadata==4.12.0
 ipython==8.5.0
-itsdangerous==2.1.2; python_version >= '3.7'
-jedi==0.18.1; python_version >= '3.6'
-jinja2==3.1.2; python_version >= '3.7'
-markupsafe==2.1.1; python_version >= '3.7'
-matplotlib-inline==0.1.6; python_version >= '3.5'
-parso==0.8.3; python_version >= '3.6'
-pexpect==4.8.0; sys_platform != 'win32'
+ipython-genutils==0.2.0
+itsdangerous==2.1.2
+jedi==0.18.1
+Jinja2==3.1.2
+Mako==1.2.2
+MarkupSafe==2.1.1
+matplotlib-inline==0.1.6
+mccabe==0.7.0
+parso==0.8.3
+pexpect==4.8.0
 pickleshare==0.7.5
 piecash==1.2.0
-prompt-toolkit==3.0.31; python_full_version >= '3.6.2'
+prompt-toolkit==3.0.31
 ptyprocess==0.7.0
 pure-eval==0.2.2
-pygments==2.13.0; python_version >= '3.6'
-pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
+pycodestyle==2.9.1
+pyflakes==2.5.0
+Pygments==2.13.0
 pytz==2022.2.1
-pyyaml==6.0
-setuptools==65.3.0; python_version >= '3.7'
-six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
-sqlalchemy-utils==0.38.3; python_version ~= '3.6'
-sqlalchemy==1.3.24; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
+pytz-deprecation-shim==0.1.0.post0
+PyYAML==6.0
+six==1.16.0
+SQLAlchemy==1.3.24
+SQLAlchemy-Utils==0.38.3
 stack-data==0.5.0
-traitlets==5.3.0; python_version >= '3.7'
-tzdata==2022.2; python_version >= '3.6'
-tzlocal==4.2; python_version >= '3.6'
+traitlets==5.4.0
+tzdata==2022.2
+tzlocal==4.2
 visitor==0.1.3
 voluptuous==0.13.1
 wcwidth==0.2.5
-werkzeug==2.2.2; python_version >= '3.7'
-wtforms==3.0.1; python_version >= '3.7'
+Werkzeug==2.2.2
+WTForms==3.0.1
+zipp==3.8.1
